requirements.txt
gitpython>=3.1.0
prompt_toolkit>=3.0.0
pyyaml>=5.1
prettytable>=3.7.0
cryptography>=41.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3
typing>=3.7.4.3
